Create hook; Put type to model folder
