+++ model/types.ts
@@ -1,0 +1,7 @@
+export type Instrument = {
+  id: number;
+  name: string;
+  type: string;
+  replacement_date?: number | null;
+  progress?: number | null;
+};
